+++ compiler/GHC/Tc/Utils/TcType.hs
@@ -2542,6 +2542,7 @@ isTerminatingClass cls
             -- Typeable constraints are bigger than they appear due
             -- to kind polymorphism, but we can never get instance divergence this way
     || cls `hasKey` coercibleTyConKey
+    || cls `hasKey` unsatisfiableClassNameKey
 
 allDistinctTyVars :: TyVarSet -> [KindOrType] -> Bool
 -- (allDistinctTyVars tvs tys) returns True if tys are